refactor(bin): move cli to bin, add dev crate, remove rundler crate
