+++ bin/tools/src/bin/deploy_dev_contracts.rs
@@ -1,7 +1,6 @@
 use ethers::utils::hex;
-use rundler_rundler::common::{
-    dev,
-    dev::{DevAddresses, BUNDLER_ACCOUNT_ID, PAYMASTER_SIGNER_ACCOUNT_ID, WALLET_OWNER_ACCOUNT_ID},
+use rundler_dev::{
+    self, DevAddresses, BUNDLER_ACCOUNT_ID, PAYMASTER_SIGNER_ACCOUNT_ID, WALLET_OWNER_ACCOUNT_ID,
 };
 
 #[tokio::main]
@@ -9,7 +8,7 @@ async fn main() -> anyhow::Result<()> {
     let bytecode = include_str!(
         "../../../../crates/types/contracts/bytecode/entrypoint/0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789.txt",
     );
-    let addresses = dev::deploy_dev_contracts(bytecode).await?;
+    let addresses = rundler_dev::deploy_dev_contracts(bytecode).await?;
     addresses.write_to_env_file()?;
 
     let DevAddresses {
@@ -25,15 +24,17 @@ async fn main() -> anyhow::Result<()> {
     println!();
     println!(
         "Bundler private key: 0x{}",
-        hex::encode(dev::test_signing_key_bytes(BUNDLER_ACCOUNT_ID))
+        hex::encode(rundler_dev::test_signing_key_bytes(BUNDLER_ACCOUNT_ID))
     );
     println!(
         "Wallet owner private key: 0x{}",
-        hex::encode(dev::test_signing_key_bytes(WALLET_OWNER_ACCOUNT_ID))
+        hex::encode(rundler_dev::test_signing_key_bytes(WALLET_OWNER_ACCOUNT_ID))
     );
     println!(
         "Paymaster private signing key: 0x{}",
-        hex::encode(dev::test_signing_key_bytes(PAYMASTER_SIGNER_ACCOUNT_ID))
+        hex::encode(rundler_dev::test_signing_key_bytes(
+            PAYMASTER_SIGNER_ACCOUNT_ID
+        ))
     );
     Ok(())
 }
