+++ bin/tools/src/bin/get_example_ops.rs
@@ -1,6 +1,6 @@
 use dotenv::dotenv;
+use rundler_dev::DevClients;
 use rundler_rpc::RpcUserOperation;
-use rundler_rundler::common::dev::DevClients;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
