+++ crates/dev/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "rundler-dev"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+rundler-types = { path = "../types" }
+
+anyhow.workspace = true
+ethers.workspace = true
