--- crates/rundler/Cargo.toml
@@ -1,69 +1,0 @@
-[package]
-name = "rundler-rundler"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-rundler-builder = { path = "../builder" }
-rundler-pool = { path = "../pool" }
-rundler-provider = { path = "../provider" }
-rundler-rpc = { path = "../rpc" } 
-rundler-sim = { path = "../sim" }
-rundler-task = { path = "../task" }
-rundler-types = { path = "../types" }
-rundler-utils = { path = "../utils" }
-
-anyhow.workspace = true
-clap = { version = "4.2.4", features = ["derive", "env"] }
-ethers.workspace = true
-futures = "0.3.28"
-futures-util = "0.3.28"
-hyper = "0.14.27"
-itertools = "0.11.0"
-metrics = "0.21.0"
-metrics-exporter-prometheus = "0.12.0"
-metrics-process = "1.0.10"
-metrics-util = "0.15.0"
-parse-display = "0.8.0"
-pin-project = "1.0.12"
-prost = "0.12.0"
-prost-types = "0.12.0"
-rand.workspace = true
-reqwest.workspace = true
-rustls = "0.21.7"
-rusoto_core = { version = "0.48.0", default-features = false, features = ["rustls"] }
-rusoto_kms = { version = "0.48.0", default-features = false, features = ["rustls"] }
-rusoto_s3 = { version = "0.48.0", default-features = false, features = ["rustls"] }
-serde = { version = "1.0.160", features = ["derive"] }
-serde_json.workspace = true
-sscanf = "0.4.0"
-strum = { version = "0.25.0", features = ["derive"] }
-thiserror.workspace = true
-tokio = { workspace = true, features = ["macros", "rt-multi-thread", "signal", "sync"] }
-tokio-rustls = "0.24.1"
-tokio-util = "0.7.8"
-tonic = "0.10.0"
-tonic-health = "0.10.0"
-tonic-reflection = "0.10.0"
-tonic-types = "0.10.0"
-tower = "0.4.13"
-tower-http = { version = "0.4.0", features = ["trace"] }
-tracing.workspace = true
-tracing-appender = "0.2.2"
-tracing-subscriber = { version = "0.3.16", features = ["env-filter", "fmt", "json"] }
-url.workspace = true
-
-[dev-dependencies]
-cargo-husky.workspace = true
-mockall.workspace = true
-rundler-pool = { path = "../pool", features = ["test-utils"] }
-rundler-provider = { path = "../provider", features = ["test-utils"] }
-rundler-sim = { path = "../sim", features = ["test-utils"] } 
-tokio-util = "0.7.7"
-
-[build-dependencies]
-ethers = "2.0.3"
-tonic-build = "0.10.0"
