--- crates/rundler/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod cli;
-pub mod common;
