+++ crates/sim/src/simulation/simulation.rs
@@ -728,7 +728,7 @@ impl Settings {
     }
 }
 
-#[cfg(feature = "test-utils")]
+#[cfg(any(test, feature = "test-utils"))]
 impl Default for Settings {
     fn default() -> Self {
         Self {
